\name{flowQA-package}
\alias{flowQA-package}
\alias{flowQA}
\docType{package}
\title{
flowQA: A package that provides automated flow data quality assessment based on gated cell populations
}
\description{
  The package provide two important methods:\code{qaCheck} and \code{plot}, which allows users to use \code{formula} as a general and flexible way
  to specify the tubes,channels,statistics and gated populations to perform differnt QA tasks.
}
\details{
\tabular{ll}{
Package: \tab flowQA\cr
Version: \tab 0.99.1\cr
Date:\tab 2011-12-02\cr
Depends: \tab R (>= 2.14), flowCore\cr
License: \tab Artistic-2.0\cr

}
}


\author{
Mike Jiang,Greg Finak

Maintainer: Mike Jiang <wjiang2@fhcrc.org>
}
\references{
\url{http://www.rglab.org/}
}
\keyword{ package }


