\docType{methods}
\name{getQAStats}
\alias{getQAStats}
\alias{getQAStats-methods}
\alias{getQAStats,environment-method}
\alias{getQAStats,GatingHierarchy-method}
\alias{getQAStats,GatingSet-method}
\alias{getQAStats,GatingSetList-method}
\title{Get statistics from gated data for QA check.}
\arguments{
  \item{obj}{A \code{\link[=GatingHierarchy-class]{gating
  hierarchy}} that stores the gated cell populations for
  one FCS file; or a
  \code{\link[=GatingSet-class]{GatingSet}} or a
  \code{\link[=GatingSetList-class]{GatingSetList}}
  containing multiple \code{gating hierarchies} or an
  \code{environment} that stores the \code{GatingSet} and
  all other information.}

  \item{...}{isFlowCore: A logical scalar indicating
  whether the statistics are the original ones in flowJo
  xml workspace or the re-calculated version by flowCore in
  R.  nslaves: An integer scalar indicating the number of
  nodes used for parallel computing when obj is
  \code{GatingHierarchy} or \code{GatingSet}.  When the
  \code{parallel} package is loaded and nslaves is NULL,
  its value is automatically decided by available number of
  nodes.  When it is set to 1, then forced to run in serial
  mode.}
}
\value{
  a data frame when obj is a \code{GatingHierarchy} or
  \code{GatingSet}
}
\description{
  \code{getQAStats} calculates and extract statistics of
  each gated cell population in the gating hierarchies
  generated by flowWorkspace package.
}
\details{
  This is the second preprocessing step followed by parsing
  gating template from flowJo workspace with
  \code{\link{parseWorkspace}}.  Different QA checks can be
  performed after this step is done.  when obj is an
  \code{environment} the results are stored as a dataframe
  with the name of "statsOfGS" in the environment.
}
\examples{
\dontrun{

getQAStats(G[[1]])#extract stats from a gating hierarchy
getQAStats(G)#from a gating set

}
}
\author{
  Mike Jiang,Greg Finak

  Maintainer: Mike Jiang <wjiang2@fhcrc.org>
}
\seealso{
  \code{\link{qaCheck}},\code{\link{qaTask}},\code{\link{qaReport}}
}
\keyword{methods}

