\name{initDB}
\alias{initDB}
\title{Initializes the data environment}
\usage{
  initDB(db = .db)
}
\arguments{
  \item{db}{An \code{environment} storing all the QA data.
  By default it is an hidden global environment
  \code{.db}.}
}
\description{
  Initializes and prepares the data environment for storing
  the QA data
}

