\name{qa.report}
\alias{qa.report}
\title{
create quality assessment report 
}
\description{
 create quality assessment report in HTML format, which contains the annotated svg summary plot for each 
 QA task.
}
\usage{
qa.report(db,outDir,splash=TRUE,plotAll=FALSE) 
}
\arguments{
  
  \item{db}{ An environment giving the data source that contains the QA results.}
  
  \item{outDir}{A character scalar giving the output path of QA report.}
    
  \item{splash}{A logical scalar indicating whether the flowQA package splash tag 'generated by \dots' should be
				written right after the report title. }
  
  \item{plotAll}{A logical scalar indicating whether to plot the 1D/2D density plot for all the individual FCS file or just
  					the FCS files that are marked as outliers by \code{\link{qaCheck}}.
  					}
}

\details{
QA results need to be calculated by calling \code{\link{qaCheck}} method before using 
this function to generate meaningful report.
It also reads the meta information of each QA task provided in spreadsheet qaCheckList and 
generate the summary tables and svg plots. Svg plots provide tooltips containing the detail information about each sample as and 
hyperlinks of densityplot for each individual FCS file.  
      
}
\author{
Mike Jiang,Greg Finak


Maintainer: Mike Jiang <wjiang2@fhcrc.org>
}

\seealso{\code{\link{qaCheck}},\code{\link[flowQA:plot]{plot}}}

\examples{

data("ITN029")#load stats from disk

qa.report(db,outDir="~/output")
}
\keyword{functions}

