\name{qaTask-class}
\docType{class}
\alias{qaTask-class}
\alias{qaTask}
\alias{qaID,qaTask}
\alias{getName,qaTask}
\alias{getName,qaTask}
\alias{description,qaTask}
\alias{qaLevel,qaTask}
\alias{getPop,qaTask}
\alias{getData,qaTask}
\alias{formula,qaTask}
\alias{plotType,qaTask}
\alias{qpar,qaTask-method}
\alias{qpar<-,qaTask,list-method}
\alias{qpar}
\alias{qpar<-}
\alias{scatterPar}
\alias{scatterPar<-}
\alias{scatterPar,qaTask-method}
\alias{scatterPar<-,qaTask,list-method}
\alias{htmlReport}
\alias{htmlReport<-}
\alias{htmlReport,qaTask-method}
\alias{htmlReport<-,qaTask,logical-method}
\alias{highlight}
\alias{highlight<-}
\alias{highlight,qaTask-method}
\alias{highlight<-,qaTask,character-method}
\alias{rFunc,qaTask-method}
\alias{rFunc}
\alias{rFunc<-}
\alias{rFunc<-,qaTask,ANY-method}
\alias{rFunc<-,qaTask-method}
\alias{show,qaTask}
\alias{makeQaTask}
\alias{read.qaTask}

\title{'qaTask': a class for storing important information of flow cytometry data quality assessment task}
\description{
This class stores the meta information, the name of gated cell population associated with the QA task and 
the formula describing how the QA task is performed.     
}

\section{Objects from the Class}{
Objects can be created by using constructor:
	makeQaTask(db=.db,qaName,description,qaLevel,pop,formula,plotType)
or 
	read.qaTask(db=.db,checkListFile)

}
\arguments{
  \item{checkListFile}{A character scalar giving the file path, which is a csv spreadsheet contains the 
  						detailed information of each QA task. It should have the columns:
  						qaID,qaName,description,qaLevel,pop,formula,plotType. See the slots of \code{\link{qaTask}}
  						for more details. }
}
  						
\section{Slots}{
  \describe{
    \item{\code{qaID}:}{A integer for the id of qaTask object }
    \item{\code{qaName}:}{A character containing the QA task name }
    \item{\code{description}:}{OA character containing the description of QA task.}
    \item{\code{qaLevel}:}{A character vector containing QA task level,
    								which is displayed in the html report.}
	\item{\code{pop}:}{A character containing the name of the cell population ,which is also equivalent to the name of the node in the gating hierarchy tree.
	The gating hierarchy is generated by flowWorkspace when the gating template is parsed from flowJo workspace into R. 
	It basically tells qaCheck or plot methods the particular gated cell population from which statistics are extracted. }

    \item{\code{formula}:}{a formula describing how the QA task is performed. 
    						see \code{\link[=qaCheck,qaTask-method]{qaCheck}} for more details. } 
     \item{\code{type}:}{a character indicate how pop is matched.}
	\item{\code{subset}:}{an expression specifying a subset of data to be applied QA task.}       
                        
    \item{\code{plotType}:}{A character indicating how the QA result is plotted. Currently only "xyplot" and "bwplot" are supported.}
    \item{\code{width}:}{A numeric scalar indicating the  width of svg figure .}
    \item{\code{height}:}{A numeric scalar indicating the  height of svg figure .}
    \item{\code{par}:}{A list storing all the lattice arguments to control the appearance of the plot. See \code{\link{xyplot}} for details.}
    \item{\code{scatterPar}:}{A list storing the arguments to control the appearance of 
    							the individual plot.
    							Example scatterPar value:
    							list(type="densityplot",scales=list(x=list(log=TRUE)))
    							this will set the log scale for x axis of the 1D-densityplot
    							}
   
    \item{\code{htmlReport}:}{A logical scalar indicating whether to plot the task when \code{\link{qaReport}} is called. 
							    By default, when there are no outliers detected for the qaTask,it is not plotted in the report.
							    But sometime it is helpful to still look at the plot with the trend of the data like MFI stability over time.
							    So by setting this flag to TRUE, we force this task to be plotted anyway. }
   \item{\code{highlight}:}{A character scalar indicating the level on which the dots in svg output should be highlighted when they are hovered over by mouse.
   							It is one of the visualization feature provided by svg format that helps to identify a couple of dots that share the same information (like FCS name,or sample ID).
   							It should be a valid column name in the metaFile (see \code{\link{getQAStats} for more details about the meta file}.
   							}
   \item{\code{rFunc}:}{A regression function passed to some qa tasks to monitor the long term trend.}
    \item{\code{db}:}{An environment containing the database connection, which stores the gating hierarchy,QA task list,sample information and outliers detection results. .}
  }
}

\value{
  
\code{read.qaTask} returns a list of qaTask objects by reading external csv spreadsheet containing descriptions of each QA task. 
  
}

	 

\author{
Mike Jiang,Greg Finak


Maintainer: Mike Jiang <wjiang2@fhcrc.org>

}


\seealso{
\code{\link{qaCheck}},\code{\link{plot}},\code{\link{qaReport}}
}

\keyword{classes}
