\name{getQAStats}
\alias{getQAStats}
\title{
Calculate and extract the statistics from the gating hierarchies.   
}
\description{
Calculate and extract statistics of each gated cell population in the gating hierarchies generated by flowWorkspace package.
The results are stored in \code{db} environment for the further quality assessment performed at different levels.
}
\usage{
getQAStats(G,metaFile) 
}
\arguments{
  
  \item{G}{ A \code{\link[flowWorkspace:GatingSet]{GatingSet}} containing multiple \code{\link[flowWorkspace:GatingHierarchy]{gating hierarchies}} within which 
  			 gated cell populations are stored.}
  
  \item{metaFile}{A character scalar giving the file path, which is a csv spreadsheet contains the sample information
  					such as subject(or patient) ID,visit number,tube(or staining panel),FCS filename.sample recording date...}
  
}

\details{
      This is the second preprocessing step followed by parsing gating template from flowJo workspace with \code{\link{parseWorkspace}}.
      Different QA checks can be performed after this step is done. 
}

\value{
  
  An environment that stores sample annotation information (\code{anno}) reading from \code{metaFile}
  ,GatingSet(\code{G}) and statistics for each gated population (\code{statsOfGS})
  
}

\author{
Mike Jiang,Greg Finak


Maintainer: Mike Jiang <wjiang2@fhcrc.org>
}

\seealso{\code{\link{qaCheck}}}

\examples{
\dontrun{
load(file="gatingHierarchy/GS.Rda")
db<-getQAStats(G,metaFile="~/FCS_File_mapping.csv")
}
}
\keyword{functions}

