\name{qaReport}
\docType{methods}
\alias{qaReport}
\alias{qaReport-methods}
\alias{qaReport,qaTask-method}
\alias{qaReport,list-method}
\title{
create quality assessment report 
}
\description{
 create quality assessment report in HTML format, which contains the annotated svg summary plot for each 
 QA task.
}
\usage{
qaReport(obj,...) 
}
\arguments{
  
  \item{obj}{ A \code{qaTask} or a list of \code{qaTask} objects }
  
 
 \item{...}{ 
		outDir:A character scalar giving the output path of QA report.

		plotAll:A logical scalar passed down to the \code{\link[QUALIFIER:plot]{plot}} method.
		
		subset: A logical expression as a filter that is passed to \code{\link{plot}}. see \code{\link{qaCheck}} for more details.  					
  		
  		splash: A logical scalar indicating whether the QUALIFIER package splash tag 'generated by \dots' should be
				written right after the report title. 
				
		title: A character scalar indicating the title of the report
		
		subTitle: A character scalar indicating  the subtitle of the report
				
		}					
}

\details{
QA results need to be calculated by calling \code{\link{qaCheck}} method before using 
this function to generate meaningful report.
It also reads the meta information of each QA task and 
generate the summary tables and svg plots. Svg plots provide tooltips containing the detail information about each sample and 
hyperlinks of densityplot for each individual FCS file.  
      
}
\author{
Mike Jiang,Greg Finak

Maintainer: Mike Jiang <wjiang2@fhcrc.org>
}

\seealso{\code{\link{qaCheck}},\code{\link[QUALIFIER:plot]{plot}}}

\examples{
\dontrun{
data("ITNQASTUDY")#load stats from disk
checkListFile<-file.path(system.file("data",package="QUALIFIER"),"qaCheckList.csv.gz")
qaTask.list<-read.qaTask(db,checkListFile)
qaReport(qaTask.list[[1]],outDir="~/output")
qaReport(qaTask.list[2:3],outDir="~/output")
}
}
\keyword{methods}

