\name{qaTask-class}
\docType{class}
\alias{qaTask-class}
\alias{qaTask}
\alias{qaID,qaTask}
\alias{getName,qaTask}
\alias{getName,qaTask}
\alias{description,qaTask}
\alias{qaLevel,qaTask}
\alias{getPop,qaTask}
\alias{getData,qaTask}
\alias{formula,qaTask}
\alias{plotType,qaTask}
\alias{qpar,qaTask-method}
\alias{qpar<-,qaTask,list-method}

\alias{show,qaTask}

\title{'qaTask': a class for storing important information of flow cytometry data quality assessment task}
\description{
This class stores the meta information, the name of gated cell population associated with the QA task and 
the formula describing how the QA task is performed.     
}
\section{Objects from the Class}{
Objects can be created by using:
\code{\link{makeQaTask}}
}
\section{Slots}{
  \describe{
    \item{\code{qaID}:}{A integer for the id of qaTask object }
    \item{\code{qaName}:}{A character containing the QA task name }
    \item{\code{description}:}{OA character containing the description of QA task.}
    \item{\code{qaLevel}:}{A character vector containing QA task level,
    								which is displayed in the html report.}
	\item{\code{pop}:}{A character containing the name of the cell population ,which is also equivalent to the name of the node in the gating hierarchy tree.
	The gating hierarchy is generated by flowWorkspace when the gating template is parsed from flowJo workspace into R. 
	It basically tells qaCheck or plot methods the particular gated cell population from which statistics are extracted. }

    \item{\code{formula}:}{a formula describing how the QA task is performed. 
    						see \code{\link[=qaCheck,qaTask-method]{qaCheck}} for more details. } 
    \item{\code{plotType}:}{A character indicating how the QA result is plotted. Currently only "xyplot" and "bwplot" are supported.}
    \item{\code{par}:}{A list storing all the lattice arguments to control the appearance of the plot. See \code{\link{xyplot}} for details.}
    \item{\code{db}:}{An environment containing the database connection, which stores the gating hierarchy,QA task list,sample information and outliers detection results. .}
  }
}

	 

\author{
Mike Jiang,Greg Finak


Maintainer: Mike Jiang <wjiang2@fhcrc.org>

}


\seealso{
\code{\link{qaCheck}},\code{\link{plot}}
}

\keyword{classes}
